joke-it/back-end/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
alt-profanity-check==1.2.2
'bitch', 'fuck', 'shit'
